requirements.txt
pandas==2.1.4
SQLAlchemy==1.4.54
psycopg2-binary==2.9.10
faker==33.1.0
requests==2.32.3
python-dotenv==1.0.1
fastapi==0.115.6
uvicorn==0.32.1
apache-airflow==2.11.1
dbt-postgres==1.8.2